requirements_llm.txt
numpy==1.21.6
PyYAML==6.0.1
python-dotenv==1.0.1
torch==2.0.0
transformers==4.43.4
accelerate==0.33.0
bitsandbytes==0.42.0
anyio==3.*
httpx==0.27.0
h11==0.13
fastapi[standard]==0.112.0
uvicorn==0.30.5
huggingface-hub==0.24.5